mem_linux.go: fix init()

+++ v3/mem_linux.go
@@ -16,5 +16,5 @@ func init() {
 		return
 	}
 
-	totalRam = si.Totalram
+	totalRam = uint64(si.Totalram)
 }
